+++ src/components/Cards/Tags/Tags.jsx
@@ -24,10 +24,10 @@ const Tags = ({ className, tags = [] }) => {
   const handleOnClick = (event, tag) => {
     event.stopPropagation();
     if (getTagInfo(tag) != null) {
-      if (location.pathname == "/testSearch") {
+      if (location.pathname == "/search") {
         location.reload();
       }
-      navigate("/testSearch", { state: { tagInfo: getTagInfo(tag) } });
+      navigate("/search", { state: { tagInfo: getTagInfo(tag) } });
     }
   };
