+++ tests/UnitTest/Spoonacular/SearchByIngridients/Tags.test.jsx
@@ -101,8 +101,8 @@ describe("Tags", () => {
 
     // use navigate is not called once?
     expect(mockUseNavigate).toBeCalledTimes(1);
-    // navigates "/testSearch" with information?
-    expect(mockUseNavigate).toBeCalledWith("/testSearch", {
+    // navigates "/search" with information?
+    expect(mockUseNavigate).toBeCalledWith("/search", {
       state: { tagInfo: getTagInfo(tags[2]) },
     });
   });
